applications/foldrun/src/analysis-job-trigger/requirements.txt
functions-framework==3.10.1
google-cloud-storage==3.10.1
google-auth==2.38.0
google-cloud-run==0.16.0
